cpu_arm: Avoid use of 'unsigned long'
Covert all use of 'unsigned long' to 'unsigned long long'.

+++ src/cpu/cpu_arm_data.h
@@ -25,7 +25,7 @@
 
 typedef struct _virCPUarmData virCPUarmData;
 struct _virCPUarmData {
-    unsigned long vendor_id;
-    unsigned long pvr;
+    unsigned long long vendor_id;
+    unsigned long long pvr;
     char **features;
 };
